geniso: initial import of the geniso tool
This tool is used to extract the files from the virtio-win image
shipped at http://alt.fedoraproject.org/pub/alt/virtio-win/latest/images/bin/
and to copy the drivers in a directory layout that the installer
can use. It makes sure to create hardlinks when the same driver
is used for several windows versions. This is useful for example if
one wants to recreate a new ISO using the generated layout without
wasting space with genisoimage --cache-inodes.