geniso: add file regex to driver definition
On the source ISO, a directory can contain files for multiple drivers.
Since a driver definition only describes a single driver, we have to
make sure we are not copying files that do not belong to the current
driver. In order to achieve this, we add a regular expression to
each driver entry describing the files that belong to the driver, and
when iterating over a directory's files, we use this regular expression
to filter the files we process.