geniso: ignore duplicate wdfcoinstaller01009.dll files
This file is used both by vioserial and the balloon driver, so
it's expected that we try to copy it twice to a given directory.
This commit adds a hack to ignore this file. We want to keep warning
for other duplicate files since apart from this file, duplicate files
are unexpected.