nsis: Add StopServices function
This method will stop vdagent and the balloon service. It's useful when
upgrading an existing installation, as otherwise we would not be able to
overwrite the running services.

This uses the SimpleSC NSIS plugin
from http://nsis.sourceforge.net/NSIS_Simple_Service_Plugin
which must be in NSIS plugin path on the machine where you build the
installer (/usr/share/nsis/Plugins/ on the fedora machine I use).