Remove EXE_VERSION
This is only used when generating the oVirt installer as it does not
want any version number to be included in the .exe filename provided on
the ISO it ships. However, if someone was to ship a standalone .exe
outside of an ISO, an unversioned installer executable is not what
is expected. This commit removes EXE_VERSION, the installer executable
will be "unversioned" at ISO time generation instead.

Change-Id: Iaa3263aa7d26f39f073db058a6138728c09e12c4
Signed-off-by: Christophe Fergeau <cfergeau@redhat.com>