build: Add ISO generation as well as make install
The preferred method of distribution of the oVirt guest tools installer
is to have it in an ISO image wrapped in an RPM. This allows to easily
share installer updates with multiple Windows guests.
This commits adds an 'iso' target to the Makefile which will generate
such an ISO which can then be consumed by the guest-tools-iso spec file.

Split off from [1].

[1] https://gerrit.ovirt.org/47771

Change-Id: I294f640d95bdcec6bc0267e79c4fed244ecbc36e
Signed-off-by: Christophe Fergeau <cfergeau@redhat.com>
Signed-off-by: Yedidyah Bar David <didi@redhat.com>